% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{build_fit_stan_data_until_week}
\alias{build_fit_stan_data_until_week}
\title{Build fit Stan data up to a specific global week from a full indexed schedule}
\usage{
build_fit_stan_data_until_week(schedule_df, last_w)
}
\arguments{
\item{schedule_df}{Output of prepare_schedule_indices()}

\item{last_w}{Global week_idx to include (<= max in schedule)}
}
\value{
Stan data list suitable for mod_fit.stan
}
\description{
Build fit Stan data up to a specific global week from a full indexed schedule
}
