% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{create_forecast_data}
\alias{create_forecast_data}
\title{Create Forecast Stan Data from Previous Fit Data}
\usage{
create_forecast_data(
  previous_stan_data,
  forecast_seasons,
  forecast_weeks = NULL,
  min_week = NULL,
  max_week = NULL,
  exclude_outcome_vars = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{previous_stan_data}{Stan data list from a previous create_stan_data() call}

\item{forecast_seasons}{Vector of seasons to include in forecast}

\item{forecast_weeks}{Vector of weeks to include (optional, defaults to all available)}

\item{min_week}{Minimum week to include in forecast}

\item{max_week}{Maximum week to include in forecast}

\item{exclude_outcome_vars}{Remove outcome variables (scores, results) for forecasting}

\item{verbose}{Print detailed information}
}
\value{
Stan data list suitable for generated_quantities
}
\description{
Creates forecast data that can be used with generated_quantities() by taking
a previous stan_data_list and extending it with forecast games. This is
specifically designed for Stan's generated_quantities block.
}
\examples{
# Forecast 2006 season using model trained on 2002-2005
forecast_data <- create_forecast_data(
  previous_stan_data = fit_stan_data,
  forecast_seasons = 2006,
  verbose = TRUE
)
}
