% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{extend_for_forecast}
\alias{extend_for_forecast}
\title{Extend Data for Forecasting}
\usage{
extend_for_forecast(base_data, forecast_weeks, verbose = FALSE)
}
\arguments{
\item{base_data}{Base game data (data frame) OR Stan data list from create_stan_data}

\item{forecast_weeks}{Number of weeks to forecast ahead}

\item{verbose}{Print extension details}
}
\description{
Intelligently extends existing game data for forecasting by detecting where
the current data ends and adding the requested number of forecast weeks.
Now uses actual season boundaries instead of hard-coding week 22.
}
\examples{
# Extend current data by 4 weeks for forecasting
extended_data <- extend_for_forecast(game_data, forecast_weeks = 4, verbose = TRUE)
}
