% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{extract_latest_states}
\alias{extract_latest_states}
\title{Extract Latest State Summary from Fit/GQ}
\usage{
extract_latest_states(fit_or_gq, teams, extract_type = "both")
}
\description{
Extracts filtered/predicted states and sigma_obs for the most recent week
}
