% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{extract_latest_states_simple}
\alias{extract_latest_states_simple}
\title{Alternative: Simple State Extraction Using Base Posterior}
\usage{
extract_latest_states_simple(fit_or_gq, teams, extract_type = "both")
}
\description{
Simpler version that uses only posterior package functions
}
