% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{find_latest_snapshot}
\alias{find_latest_snapshot}
\title{Find the latest snapshot file in a directory}
\usage{
find_latest_snapshot(path)
}
\arguments{
\item{path}{Directory containing snapshot_wXXX.rds files}
}
\value{
Full file path of the latest snapshot or NULL if none
}
\description{
Find the latest snapshot file in a directory
}
