% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{get_last_fitted_season}
\alias{get_last_fitted_season}
\title{Last fitted season index from Stan data list}
\usage{
get_last_fitted_season(fit_stan_data)
}
\arguments{
\item{fit_stan_data}{List created by create_stan_data() used for fitting}
}
\value{
Integer last season index included in the fit
}
\description{
Last fitted season index from Stan data list
}
