% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{get_last_fitted_week}
\alias{get_last_fitted_week}
\title{Last fitted global week from Stan data list}
\usage{
get_last_fitted_week(fit_stan_data)
}
\arguments{
\item{fit_stan_data}{List created by create_stan_data() used for fitting}
}
\value{
Integer last global week index included in the fit
}
\description{
Last fitted global week from Stan data list
}
