% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{lookup_week_idx}
\alias{lookup_week_idx}
\title{Lookup global week_idx for a given season/week pair}
\usage{
lookup_week_idx(schedule_df, season, week)
}
\arguments{
\item{schedule_df}{Output of prepare_schedule_indices()}

\item{season}{Numeric season (e.g. 2006)}

\item{week}{Numeric week within season (e.g. 1)}
}
\value{
Integer global week_idx used in Stan data
}
\description{
Lookup global week_idx for a given season/week pair
}
