% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{next_week_targets}
\alias{next_week_targets}
\title{Compute next global week targets automatically (week_idx)}
\usage{
next_week_targets(fit_stan_data, horizon = 1L)
}
\arguments{
\item{fit_stan_data}{Training data list used in fit}

\item{horizon}{Integer number of weeks ahead (default 1)}
}
\value{
Integer vector of global week_idx targets (last_w + 1 ... + horizon)
}
\description{
Compute next global week targets automatically (week_idx)
}
