% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflytics-package.R
\docType{package}
\name{nflytics-package}
\alias{nflytics-package}
\alias{nflytics}
\title{nflytics: Bayesian NFL team strength tools}
\description{
Provides data preparation helpers, Stan models, and posterior summarisation
utilities for estimating latent team strengths and forecasts for NFL games.
}
\author{
\strong{Maintainer}: Tyler Pollard \email{TylerPollard410@gmail.com}

}
\keyword{internal}
