% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{prepare_gq_data}
\alias{prepare_gq_data}
\title{Prepare full GQ data list}
\usage{
prepare_gq_data(fit_stan_data, schedule_df, targets)
}
\arguments{
\item{fit_stan_data}{Training data list used in fit}

\item{schedule_df}{Full schedule with indices (see prepare_schedule_indices)}

\item{targets}{Integer vector of global week_idx to score (can exceed N_weeks)}
}
\value{
Stan data list for generate_quantities()
}
\description{
Prepare full GQ data list
}
