% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{prepare_schedule_indices}
\alias{prepare_schedule_indices}
\title{Prepare a schedule with indices consistent with Stan}
\usage{
prepare_schedule_indices(seasons = all_seasons, teams)
}
\arguments{
\item{teams}{Character vector of team abbreviations (same used for create_stan_data)}

\item{games_df}{Data frame from load_game_data()}
}
\value{
Data frame with home_idx, away_idx, season_idx, week_idx, hfa, fw/lw flags
}
\description{
Prepare a schedule with indices consistent with Stan
}
