% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_strength_api.R
\name{print.nflytics_gq}
\alias{print.nflytics_gq}
\title{Print method for latent strength generated quantities}
\usage{
\method{print}{nflytics_gq}(x, ...)
}
\description{
Print method for latent strength generated quantities
}
