% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{reuse_metric_from_fit}
\alias{reuse_metric_from_fit}
\title{Extract metric info for reuse from previous fit}
\usage{
reuse_metric_from_fit(fit)
}
\description{
Extract metric info for reuse from previous fit
}
