% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{roll_forward_fit_stan_data}
\alias{roll_forward_fit_stan_data}
\title{Roll forward fit Stan data by a number of global weeks (no OOS)}
\usage{
roll_forward_fit_stan_data(fit_stan_data, schedule_df, weeks_ahead = 1L)
}
\arguments{
\item{fit_stan_data}{List used for previous fit}

\item{schedule_df}{Data frame from prepare_schedule_indices() (must include week_idx, season_idx, home_idx, away_idx, hfa and outcomes if used)}

\item{weeks_ahead}{Integer number of global weeks to extend (default 1)}
}
\value{
Stan data list suitable for mod_fit.stan
}
\description{
Builds a new Stan fit data list including all games with week_idx <= last_w + weeks_ahead.
Uses the same variable names expected by mod_fit.stan. Outcomes must exist for all included games.
}
