% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{run_gq}
\alias{run_gq}
\title{Convenience wrapper to run generated quantities}
\usage{
run_gq(
  gq_mod,
  fit,
  gq_data,
  seed = 52,
  sig_figs = 10,
  parallel_chains = min(4L, parallel::detectCores())
)
}
\arguments{
\item{gq_mod}{Compiled cmdstanr model for mod_gq.stan}

\item{fit}{Fitted object from mod_fit.stan (cmdstanr fit)}

\item{gq_data}{Data list from prepare_gq_data()}

\item{seed}{RNG seed}

\item{sig_figs}{Integer}

\item{parallel_chains}{Integer}
}
\value{
cmdstanr gq fit object
}
\description{
Convenience wrapper to run generated quantities
}
