% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{summarise_backtest_results}
\alias{summarise_backtest_results}
\title{Summarise sequential backtest diagnostics}
\usage{
summarise_backtest_results(results_list)
}
\arguments{
\item{results_list}{List element returned by run_sequential_backtest()$results}
}
\value{
List with tibble of OOS predictions and tibble of aggregate metrics
}
\description{
Summarise sequential backtest diagnostics
}
